Abstraction of various drivers
